+++ packages/sdk/CHANGELOG.md
@@ -11,7 +11,8 @@ commit that affects the code in a way that consumers might care about should inc
 
 ## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.48...HEAD)
 
-No changes.
+-   Adds optional `renderInvalidCellValue` prop to `RecordCard` and `CellRenderer` to render a
+    component if validation fails.
 
 ## [0.0.48](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.47...@airtable/blocks@0.0.48) - 2020-04-30